Template src must be specified explicitly via CLI
Before, we accepted a template name and then performed a sort of
heuristic to determine if it was an alias, git URL, or local filesystem
path. This was a bit too magical.

Now, the template must be specified with either the `-t`, `r`, or `p`
option, for template alias, repository, or path, respectively. We use
yargs-parser to parse the CLI options and then delegate to the
appropriate template resolver.

The template resolvers now throw errors if they can't find the template
that was specified, rather than returning `undefined`. This gives the
user more explicit and actionable error messages.

Also add more test coverage for each of the resolvers.