Upgrade CircleCI config to use node orb v4
According to the example provided by CircleCI, the way to install
and cache packages via yarn is to use this new command:

```
- node/install-packages:
    pkg-manager: yarn
```

https://github.com/CircleCI-Public/node-orb/blob/v4.1.0/src/examples/run_tests_with_yarn.yml

The orb will automatically cache and restore the `node_modules`
directory between builds.

That means that a separate `install_deps` job is not really necessary.
It solves for a race condition where two jobs running in parallel both
have a dirty `node_modules` cache and both have to install new versions
of dependencies. In practice, though, dependencies are not changed all
that often, so the extra `install_deps` prerequisite job only slows down
the common scenario where the cache does not need to change.

Note that behind the scenes the `node/default` executor is now using the
new `cimg/node` docker image.

https://github.com/CircleCI-Public/cimg-node